New Blog Feature
Added a new php page that is going to be the beginning of a new blog
project. In it's current state, it reads from a mysql database and
posts all "posts" to the page.

The backend is in a semi-functional state, so new developments will
mostly be on the HTML/CSS side.