Footer with contact info
Added the first draft of a footer. On the left it displays a notice
that the site is open-sourced on github and then on the right it
displays an inline list of contact sources, including email, twitter,
and linkedin.

While all buttons are functional, they don't look very pretty,
so I'll have to look into making it all look a little more uniform and nice.